requirements.txt
# pandas==2.0.3
# mysql-connector-python==8.1.0
# python-dotenv==1.0.0
# sentence-transformers==2.2.2
# qdrant-client==1.7.0
# openpyxl==3.1.2
# PyPDF2==3.0.1
# requests==2.31.0

pandas==2.0.3
mysql-connector-python==8.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
sentence-transformers==2.7.0
transformers==4.30.0
qdrant-client==1.9.1
openpyxl==3.1.2
PyPDF2==3.0.1
requests==2.31.0
protobuf==3.20.3
torch==2.0.1
numpy==1.24.3

# New packages for Phase 2
streamlit==1.28.1
# langchain==0.1.0
# langchain-community==0.0.10
# openai==1.6.1
matplotlib==3.7.1
plotly==5.18.0
reportlab==4.0.7
pillow==10.0.0